DDF-6565 create a video thumbnail service
