+++ catalog/common/catalog-common-video-thumbnail-impl/src/test/resources/corrupted.mp4
