+++ catalog/common/catalog-common-video-thumbnail-impl/src/test/resources/long.mp4
