+++ catalog/common/catalog-common-video-thumbnail-impl/src/test/resources/medium.mp4
