+++ catalog/common/catalog-common-video-thumbnail-impl/src/test/resources/short.mp4
